requirement.txt
beautifulsoup4==4.9.3
bs4==0.0.1
celery==5.0.5
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-autoslug==1.9.8
django-cache-url==3.2.3
django-compressor==2.4
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-filter==2.4.0
django-haystack==3.0
django-imagekit==4.0.2
django-ipware==3.0.2
django-mdeditor==0.1.18
django-uuslug==1.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docutils==0.16
drf-jwt==1.17.3
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
PyMySQL==0.10.1
pyparsing==2.4.7
python-markdown-math==0.6
python-memcached==1.59
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
WeRoBot==1.13.1
Whoosh==2.7.4